Refactored InlineNodeComponent
