+++ packages/inline-node/_inline-node.css
@@ -2,11 +2,10 @@
   cursor: default;
   /* webkit disables selection for draggable elements */
   -webkit-user-select: text;
+  position: relative;
 }
 
-.sc-inline-node > .se-container { position: relative; }
-.sc-inline-node > .se-container > .se-blocker {
-  position: absolute; top:0; bottom: 0; left: 0; right: 0;
+.sc-inline-node > .se-blocker {
+  cursor: default;
+  position: absolute; left:0;
 }
-
-.sc-inline-node > .se-slug { opacity: 0; }
